requirements.txt
requests
beautifulsoup4
json5
pandas
streamlit
plotly
pyodbc
azure-functions
